requirements.txt
python-dotenv
google-genai
flask